fix all warnings on default level

+++ src/board_controller/openbci/ganglion_bglib/main.cpp
@@ -102,7 +102,7 @@ namespace GanglionLib
         state = State::OPEN_CALLED;
         char *mac_addr = (char *)param;
         // convert string mac addr to bd_addr struct
-        for (int i = 0; i < strlen (mac_addr); i++)
+        for (int i = 0; i < (int)strlen (mac_addr); i++)
         {
             mac_addr[i] = tolower (mac_addr[i]);
         }
